BatchStateMapper
