+++ source/dotnet/Services/Domain/BatchExecutionStateDomainService/BatchStateMapper.cs
@@ -1,0 +1,34 @@
+﻿// Copyright 2020 Energinet DataHub A/S
+//
+// Licensed under the Apache License, Version 2.0 (the "License2");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//     http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+using Energinet.DataHub.Wholesale.Domain.BatchAggregate;
+using Energinet.DataHub.Wholesale.Domain.CalculationDomainService;
+
+namespace Energinet.DataHub.Wholesale.Domain.BatchExecutionStateDomainService;
+
+public class BatchStateMapper : IBatchStateMapper
+{
+    public BatchExecutionState MapState(CalculationState calculationState)
+    {
+        return calculationState switch
+        {
+            CalculationState.Pending => BatchExecutionState.Pending,
+            CalculationState.Running => BatchExecutionState.Executing,
+            CalculationState.Completed => BatchExecutionState.Completed,
+            CalculationState.Canceled => BatchExecutionState.Canceled,
+            CalculationState.Failed => BatchExecutionState.Failed,
+            _ => throw new ArgumentOutOfRangeException(nameof(calculationState), calculationState, "Unexpected JobState."),
+        };
+    }
+}
